Restored HLS support for 64bit ReconOS, adjusted sort_demo accordingly

+++ lib/thread/reconos_thread_hls.h
@@ -1,6 +1,6 @@
 /*
  *                                                        ____  _____
- *                            ________  _________  ____  / __ \/ ___/
+ *                            ________  _________  ____  / __ \/ ___/64
  *                           / ___/ _ \/ ___/ __ \/ __ \/ / / /\__ \
  *                          / /  /  __/ /__/ /_/ / / / / /_/ /___/ /
  *                         /_/   \___/\___/\____/_/ /_/\____//____/
@@ -49,9 +49,9 @@
  *   }
  }
  */
-#define THREAD_ENTRY() void rt_imp(hls::stream<uint32_t> osif_sw2hw,\
-                                   hls::stream<uint32_t> osif_hw2sw,\
-                                   hls::stream<uint32_t> memif_hwt2mem,\
-                                   hls::stream<uint32_t> memif_mem2hwt)
+#define THREAD_ENTRY() void rt_imp(hls::stream<uint64_t> osif_sw2hw,\
+                                   hls::stream<uint64_t> osif_hw2sw,\
+                                   hls::stream<uint64_t> memif_hwt2mem,\
+                                   hls::stream<uint64_t> memif_mem2hwt)
 
 #endif /* RECONOS_THREAD_H */
